feat: implement react navigation
