+++ navigation/navigation.tsx
@@ -1,0 +1,15 @@
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { RootStackParamList } from "../types/navigation.types";
+import SigninScreen from "../screens/SigninScreen";
+import TestScreen from "../screens/TestScreen";
+
+export default function RootStack() {
+  const Stack = createNativeStackNavigator<RootStackParamList>();
+
+  return (
+    <Stack.Navigator screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="SigninScreen" component={SigninScreen} />
+      <Stack.Screen name="Test" component={TestScreen} />
+    </Stack.Navigator>
+  );
+}
