+++ types/navigation.types.ts
@@ -1,0 +1,4 @@
+export type RootStackParamList = {
+  SigninScreen: undefined;
+  Test: undefined;
+};
